+++ backend/config/urls.py
@@ -51,7 +51,7 @@ urlpatterns = [
     # React views
     path('', render_react_view, {'component_name': 'IndexView'}),
     path('all_view/', render_react_view, {'component_name': 'AllView'}),
-    path('adventure/', render_react_view, {'component_name': 'chooseAdventureView'}),
+    path('adventure/', render_react_view, {'component_name': 'ChooseAdventureView'}),
     edx_path('map_quiz/', 'MapQuiz'),
     edx_path('budget_voting_simulation/', 'BudgetVotingSimViz'),
     edx_path('heat_map/', 'DemocracyViz'),
