+++ frontend/src/app.js
@@ -13,7 +13,7 @@ import { MapQuiz } from './mapQuiz/mapQuiz.js';
 import { BudgetVotingSimViz } from './budgetSimulation/budgetSimulation.js';
 import { DemocracyViz } from './democracyViz/democracyViz.js';
 import { EdXView } from "./UILibrary/components";
-import { chooseAdventureView } from "./chooseAdventure/chooseAdventure.js";
+import { ChooseAdventureView } from "./chooseAdventure/chooseAdventure.js";
 
 // Import all styles
 import './UILibrary/styles.scss';
@@ -29,5 +29,5 @@ window.app_modules = {
     BudgetVotingSimViz,
     DemocracyViz,
     EdXView,
-    chooseAdventureView,
+    ChooseAdventureView,
 };
