+++ frontend/src/chooseAdventure/chooseAdventure.js
@@ -1,59 +1,27 @@
 import React from 'react';
 // import * as PropTypes from 'prop-types';
-
-// TODO: hardcode data structure here
-
-const MEDIA_STAGE = {
-    'text': 'You chose to start a media campaign.',
-    'options': [{
-        'text': 'Twitter',
-    },
-    {
-        'text': 'Facebook',
-    },
-    {
-        'text': 'Radio',
-    }],
-};
-
-const DIRECT_STAGE = {
-    'text': 'You chose to take direct action against the school.',
-    'options': [{
-        'text': 'Sue the principal',
-    },
-    {
-        'text': 'Ask the principal nicely',
-    }],
-};
-
-const START_STAGE = {
-    'text': 'Your school district\'s budget was cut!',
-    'options': [{
-        'text': 'Start a media campaign',
-        'stage': MEDIA_STAGE,
-    },
-    {
-        'text': 'Take direct action',
-        'stage': DIRECT_STAGE,
-    }],
-};
+// import { IntroView } from './introView.js';
 
 
 /**
  * Component for displaying choose your own adventure skeleton
  */
 
-export class chooseAdventureView extends React.Component {
+export class ChooseAdventureView extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            stage: START_STAGE,
+            view: 'intro',
         };
     }
 
     render() {
 
-        return ( <div>hello</div> );
+        return (
+            <div>
+                {this.state.view === 'intro' && "hello"}
+            </div>
+        );
 
     }
 }
